apis/gemini-response/requirements.txt
python-dotenv
flask
google-generativeai